refactor(admin): split problem list modules
Why:
- keep the problems route imports stable while separating list rendering from filter and pagination link logic
- reduce the mixed problems-sections module into focused route pieces

Verify:
- pnpm exec prettier --check apps/admin/src/app/quantercise/problems/page.tsx apps/admin/src/app/quantercise/problems/problems-content.tsx apps/admin/src/app/quantercise/problems/problems-sections.tsx apps/admin/src/app/quantercise/problems/problems-filters.tsx apps/admin/src/app/quantercise/problems/problems-list.tsx
- git diff --check -- apps/admin/src/app/quantercise/problems/page.tsx apps/admin/src/app/quantercise/problems/problems-content.tsx apps/admin/src/app/quantercise/problems/problems-sections.tsx apps/admin/src/app/quantercise/problems/problems-filters.tsx apps/admin/src/app/quantercise/problems/problems-list.tsx
- pnpm --filter @anipotts/admin typecheck
- pnpm --filter @anipotts/admin lint
- pnpm --filter @anipotts/admin build
- pnpm validate
